refactor: Loop through snapshot tests for each theme
