+++ src/components/Alert/index.test.tsx
@@ -9,24 +9,17 @@
 
 import * as React from 'react';
 
-import render from '~/helpers/test/render';
+import { render, THEMES } from '~/helpers/test';
 
 import { Alert } from '.';
 
 describe('Alert', () => {
-  it('ACT theme matches the snapshot', () => {
+  test.each(THEMES)('%s theme matches the snapshot', theme => {
     const { container } = render(
       <Alert text="Some alert description goes here." title="Title Here" />,
-      'ACT',
+      theme,
     );
-    expect(container).toMatchSnapshot();
-  });
 
-  it('ACT_ET theme matches the snapshot', () => {
-    const { container } = render(
-      <Alert text="Some alert description goes here." title="Title Here" />,
-      'ACT_ET',
-    );
     expect(container).toMatchSnapshot();
   });
 });
