+++ src/components/Button/index.test.tsx
@@ -9,24 +9,17 @@
 
 import * as React from 'react';
 
-import render from '~/helpers/test/render';
+import { render, THEMES } from '~/helpers/test';
 
 import { Button } from '.';
 
 describe('Button', () => {
-  it('ACT theme matches the snapshot', () => {
+  test.each(THEMES)('%s theme matches the snapshot', theme => {
     const { container } = render(
       <Button color="primary">Click me</Button>,
-      'ACT',
+      theme,
     );
-    expect(container).toMatchSnapshot();
-  });
 
-  it('ACT_ET theme matches the snapshot', () => {
-    const { container } = render(
-      <Button color="primary">Click me</Button>,
-      'ACT_ET',
-    );
     expect(container).toMatchSnapshot();
   });
 });
