+++ src/components/Checkbox/index.test.tsx
@@ -9,21 +9,14 @@
 
 import * as React from 'react';
 
-import render from '~/helpers/test/render';
+import { render, THEMES } from '~/helpers/test';
 
 import { Checkbox } from '.';
 
 describe('Checkbox', () => {
-  it('ACT theme matches the snapshot', () => {
-    const { container } = render(<Checkbox checked color="primary" />, 'ACT');
-    expect(container).toMatchSnapshot();
-  });
+  test.each(THEMES)('%s theme matches the snapshot', theme => {
+    const { container } = render(<Checkbox checked color="primary" />, theme);
 
-  it('ACT_ET theme matches the snapshot', () => {
-    const { container } = render(
-      <Checkbox checked color="primary" />,
-      'ACT_ET',
-    );
     expect(container).toMatchSnapshot();
   });
 });
