+++ src/components/Container/index.test.tsx
@@ -9,18 +9,14 @@
 
 import * as React from 'react';
 
-import render from '~/helpers/test/render';
+import { render, THEMES } from '~/helpers/test';
 
 import { Container } from '.';
 
 describe('Container', () => {
-  it('ACT theme matches the snapshot', () => {
-    const { container } = render(<Container>Hey</Container>, 'ACT');
-    expect(container).toMatchSnapshot();
-  });
+  test.each(THEMES)('%s theme matches the snapshot', theme => {
+    const { container } = render(<Container>children</Container>, theme);
 
-  it('ACT_ET theme matches the snapshot', () => {
-    const { container } = render(<Container>Hey</Container>, 'ACT_ET');
     expect(container).toMatchSnapshot();
   });
 });
